Applies a buffer rule based on the creation of a buffer zone for illiquid stocks

This rule is applied to reduce portfolio turnover and transaction costs.

Usage

apply_turnover_cap_rule(
  stock_universe_m_d_ref,
  eligibility_quantile_range,
  quantile_range_buffer,
  updated_port_weights_m_lstd_ref,
  liquidity_m_d_ref,
  liquidity_floor_cutoffs,
  turnover_cap_rule,
  verbose = TRUE
)

Arguments

stock_universe_m_d_ref

A data frame of selected signals generated by the generate_signal_matrix function.

eligibility_quantile_range

Numeric vector of length 2 indicating the quantile range for the stock universe.

quantile_range_buffer

Numeric value indicating the quantile valid for buffer zones.

updated_port_weights_m_lstd_ref

A data frame containing columns for id, tickers, dates and weights from the old portfolio (pre-rebalancing) All tickers in current stock universe must have a unique correspondence in this data frame.

liquidity_m_d_ref

A data frame (similar to features_m_df) containing columns for id, tickers, dates, and one or more market liquidity measures (e.g., inflation-adjusted mean financial volume). All tickers in current stock universe must have a unique correspondence in this data frame.

liquidity_floor_cutoffs

A list of named vectors containing cutoff values for liquidity metrics specified in liquidity_m_d_ref. The names should match the metrics and values should be the minimum acceptable values (adjust for inflation) Elements should be: "micro_caps", "small_caps", "mid_caps" and "large_caps" Classification should be in ascending order (from lest liquid to most liquid) for all metrics.

turnover_cap_rule

Character string specifying the buffer rule target (one of "micro_caps", "small_caps", "mid_caps", "large_caps", "mega_caps"). Stocks that are less liquid than turnover_cap_rule and that are included in the buffer_zone will be promoted. Stocks meeting this rule will be included in the filtered_stock_universe even if they fail the liquidity floor rule.

verbose

Logical indicating whether to print messages to the console.