Create a cap_weighted portfolio

Create a cap_weighted portfolio

Usage

create_cap_scaled_portfolio(
  universe_m_d_ref,
  liquidity_m_d_ref,
  cap_weighting_metric,
  lower_quantile_winsorization = 0.025,
  upper_quantile_winsorization = 0.975,
  verbose = TRUE
)

Arguments

universe_m_d_ref

A dataframe with tickers, is_stock_eligible and final_signal columns generated by filter_stock_universe

liquidity_m_d_ref

A data frame (similar to features_m_df) containing columns for id, tickers, dates, and one or more market liquidity measures (e.g., inflation-adjusted mean financial volume). All tickers in current stock universe must have a unique correspondence in this data frame.

cap_weighting_metric

cap weighting metric in liquidity_m_d_ref to be used to weight stocks when portfolio_construction_method is CW

lower_quantile_winsorization

lower quantile to winsorize cap_weighting_metric

upper_quantile_winsorization

upper quantile to winsorize cap_weighting_metric

verbose

If TRUE, will print messages to the console