Create individual box constraints for MVO portfolios based on liquidity, concentration and turnover constraints

Create individual box constraints for MVO portfolios based on liquidity, concentration and turnover constraints

Usage

generate_box_constraints(
  universe_m_d_ref,
  liquidity_constraint_policy = NULL,
  turnover_constraint_policy = NULL,
  concentration_constraint_policy,
  verbose = TRUE
)

Arguments

universe_m_d_ref

A dataframe of assets with tickers, is_eligible, final_signal columns and also those generated by application of constraints policies in classify_investment_universe

liquidity_constraint_policy

Optional. A named list containing objects used to apply liquidity constraints. Possible elements of the list are:

liquidity_floor_rule: A character indicating the liquidity classification (e.g., micro_caps, small_caps) used to filter stocks. Stocks with less liquidity than specified in liquidity_floor_rule will be considered ineligible. In the case of the generate_box_constraints function, liquidity_constraint_policy can also contain:

liquidity_cap_rule lists: One or many lists used to create upper bounds for weights based on a liquidity classification. Each list must contain:

liquidity_classification: A character indicating the classification for the cap.

liquidity_cap: A numeric value indicating the cap (upper bound) for stocks with that liquidity classification. Many liquidity caps might be created, and in this case, each liquidity_cap_rule must be identified with a number (e.g., liquidity_cap_rule_1, liquidity_cap_rule_2, and so on).

turnover_constraint_policy

A named list containing objects used to build buffer zones and apply turnover constraints.

Each element will constitute a buffer_zone, being a list with three elements:

liquidity_classification element: A liquidity classification (e.g., "micro_caps", "small_caps") for that buffer zone.

top_stock_quantile_buffer: A numeric value indicating a buffer value that relaxes top_stocks_quantile for stocks with the specified liquidity classification.

turnover_cap: A numeric value specifying the turnover cap. Stocks that are less liquid than specified for a buffer zone and have a signal higher than the respective buffer quantile will be considered eligible, even if they do not meet the liquidity_floor_rule.

concentration_constraint_policy

A named list containing up to four elements:

benchmark: A character vector describing the benchmark to be used to apply constraint. Must have a correspondence in benchmark_weights_m_d_ref

max_abs_active_individual_weight: The maximum absolute individual active weights.

max_abs_active_sector_weight: The maximum absolute sector active weight used for creating group constraints in generate_sector_constraints. If a given sector has no eligible stock, the one with the greatest signal will be automatically promoted.

verbose

Logical. If TRUE, the function will print messages indicating the progress of the process. Note that, in the context of generate_sector_constraints, a benchmark_weights_m_d_ref data frame must also be supplied.