Create group constraints for MTO portfolios

Create group constraints for MTO portfolios

Usage

generate_group_constraints(
  universe_m_d_ref,
  concentration_constraint_policy,
  groups_m_d_ref,
  verbose = TRUE
)

Arguments

universe_m_d_ref

A dataframe of assets with tickers, is_eligible, final_signal columns and also those generated by application of constraints policies in filter_stock_universe

concentration_constraint_policy

A named list containing up to four elements:

benchmark: A character vector describing the benchmark to be used to apply constraint. Must have a correspondence in benchmark_weights_m_d_ref

max_abs_active_individual_weight: The maximum absolute individual active weights.

max_abs_active_sector_weight: The maximum absolute sector active weight used for creating group constraints in generate_sector_constraints. If a given sector has no eligible stock, the one with the greatest signal will be automatically promoted. Note that, in the context of generate_sector_constraints, a benchmark_weights_m_d_ref data frame must also be supplied.

groups_m_d_ref

A data frame containing columns for id, tickers, dates, and sector classification columns following a given classification method. All tickers in the current stock universe must have a unique correspondence in the data frame.

verbose

If TRUE, will print messages to the console